requirements.txt
aiofiles==25.1.0
aiosqlite==0.22.1
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
asgi-lifespan==2.1.0
attrs==25.4.0
bcrypt==4.0.1
beanie==2.0.1
black==25.12.0
boto3==1.42.5
botocore==1.42.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpie==3.2.4
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jmespath==1.0.1
jq==1.10.0
lazy-model==0.4.0
librt==0.7.3
limits==5.8.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.1
multidict==6.7.0
mypy==1.19.0
mypy_extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.16.0
pypdfium2==5.5.0
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
selenium==4.41.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
starlette==0.37.2
trio==0.33.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.3
uv==0.10.2
uvicorn==0.25.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
wsproto==1.3.2
